Disable proxy by default.
By default in firefox, it is configured to use the system proxy. In
general, this is no proxy anyways, unless you have configured it.

In Windows, one way to break IE is to make a Windows system proxy to go
to back to localhost so it can't talk to the outside world.

Making such a change would then break firefox leaving at the default
setting. This setting will let firefox continue to work as normal.