Disabled speculative connections in search bar.
Speculative connections are a predictive feature, where Firefox tries to
predict the likelihood of you entering one of the search suggestions. If
it thinks you will, it starts making a connection to that search result
before you actually enter it in order to create the illusion that it is
faster, even though this speedup is quite negligible.