Added an entry for DNT Headers
While I leave it disabled, I thought I would at least add the item into
the list to explicitly set that. Also, I need to set the noscript
setting for DNT.

DNT is not often honoured by websites, and no legitimate website will
honour it anyways. It can however be used to make it easier to track you
which is a little ironic.